implement order place -- errors and failures are not handled yet !

+++ backend/order/src/main/java/com/ecom/order/cart/CartDetails.java
@@ -16,7 +16,7 @@ public record CartDetails(
     public record CartItem(
         Long productId,
         int quantity,
-        double price
+        BigDecimal price
     ) {
     }
